Home Page Updates, Added Lists, Added DB
Added the lists and create lists within the home page. Users can select items from a database of groceries and add it to their lists. They can edit the name the list and the contents.

Working on groceries DB within Firebase that allows users to search for grocery items without having to type the full thing.